+++ lua/render-markdown/types.lua
@@ -193,7 +193,7 @@
 ---@field highlight? string
 
 ---@class (exact) render.md.latex.UserConfig: render.md.base.UserConfig
----@field converter? string
+---@field converter? string|string[]
 ---@field highlight? string
 ---@field position? render.md.latex.Position
 ---@field top_pad? integer