+++ tests/helpers/system.lua
@@ -25,7 +25,7 @@ end
 local M = {}
 
 ---@param command string
----@param outputs table<string, string>
+---@param outputs table<string, string[]>
 function M.mock(command, outputs)
     stub.new(vim.fn, 'executable', function(expr)
         assert.same(command, expr)
@@ -35,7 +35,7 @@ function M.mock(command, outputs)
         assert.same({ command }, cmd)
         local output = outputs[opts.stdin]
         assert.not_nil(output, ('missing output: %s'):format(opts.stdin))
-        return Task.new(output)
+        return Task.new(table.concat(output, '\n') .. '\n')
     end)
 end